+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [Unreleased]
+
+### Added
+
+- Added latency reporting
+
 ## [1.15.0] - 2024-09-17
 
 ### Added